Issue 86 - Refactor plugin to enable via 'use' method
