+++ lib/mentionsPlugin.js
@@ -51,6 +51,11 @@ function isEmail (prefix, mention, suffix) {
   return prefix && suffix && !mention.includes('_');
 }
 
+/**
+ * Checks the character at `state.src[state.pos]` to see if it's an @
+ * If that's the case and it is not part of a link or an e-mail address it
+ * creates a `mention` token that can be rendered by `renderMentions`
+ */
 function parseMentions (state, silent) {
   if (state.src.charCodeAt(state.pos) !== 0x40 /* @ */) {
     return false;
@@ -76,4 +81,22 @@ function parseMentions (state, silent) {
   return true;
 }
 
-module.exports = parseMentions;
+/**
+ * Renders `mention` tokens into html spans with optional classes dependent on
+ * the `userName` and `displayName`
+ */
+function renderMentions (tokens, id, _, env) {
+  var mention = tokens[id].content;
+  var mentionedText = mention.slice(1); // Remove @
+  var isCurrentUserMention = mentionedText === env.userName || mentionedText === env.displayName;
+  var highlight = isCurrentUserMention ? ' at-highlight' : '';
+
+  return '<span class="at-text' + highlight + '">' + mention + '</span>';
+}
+
+function mentionsPlugin (md) {
+  md.inline.ruler.push('mentions', parseMentions);
+  md.renderer.rules.mention = renderMentions;
+}
+
+module.exports = mentionsPlugin;
